# fixed time drift by using a self-adjusting timer