No more imprecise timing! (unless you're cpu-bound or alt-tabbed or something)

Timing is now correct to +/- 3ms. Before, it would sometimes drift to at least +3 seconds per hour.

[Before, 60s](https://i.imgur.com/tnFVJes.png)
[After, 60s](https://i.imgur.com/xa7458d.png)

Even long tests are now free of time drift:
![hour long test, after](https://i.imgur.com/5bysYvS.png)

This should also result in more accurate stats.